pyproject.toml
[tool.poetry]
name = "innoscream"
version = "0.1.0"
description = "QAce team InnoScream project for SQRS course."
authors = [
    "Nikita Sannikov <n.sannikov@innopolis.university>",
    "Ilia Sardanadze <i.sardanadze@innopolis.university>",
    "Dmitriy Ortyakov <d.ortyakov@innopolis.university>",
    "Yaroslav Prudnikov <y.prudnikov@innopolis.university>",
    "Ezekiel John Gadzama <e.gadzama@innopolis.university>"
]
readme = "README.md"
packages = [{include = "innoscream", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"  # <-- This is the ONLY place Python version should be specified
fastapi = { version = ">=0.115.12,<0.116.0", extras = ["all"] }
aiogram = { version = ">=3.0.0.post0,<4.0.0", extras = ["contrib"] }
aiosqlite = ">=0.21.0,<0.22.0"
httpx = ">=0.28.1,<0.29.0"
pydantic = ">=2.11.3,<3.0.0"
python-dotenv = ">=1.1.0,<2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
flake8 = "^7.2.0"
bandit = "^1.8.3"
pydocstyle = "^6.3.0"
radon = "^6.0.1"
mutmut = "^3.2.3"
locust = "^2.35.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[tool.coverage.run]
source = ["src"]
omit = ["*/__init__.py"]

[tool.coverage.report]
show_missing = true
fail_under = 60

[tool.flake8]
max-line-length = 120
exclude = ["__init__.py", "migrations"]
max-complexity = 10